requirements.txt
python-telegram-bot[webhooks]==20.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pytesseract==0.3.10
pillow==10.0.0
opencv-python-headless==4.8.0.74
numpy==1.24.3